simplify test file
