+++ onnxruntime/python/tools/kernel_explorer/kernels/gemm_test.py
@@ -9,7 +9,7 @@ from itertools import product
 import kernel_explorer as ke
 import numpy as np
 import pytest
-from utils import get_gemm_bound
+from utils import get_basic_size, get_bert_sizes, get_gemm_bound, transab_to_suffix
 
 
 def dtype_to_suffix(dtype):
@@ -17,15 +17,6 @@ def dtype_to_suffix(dtype):
         "float32": "float",
         "float16": "half",
     }[dtype]
-
-
-def transab_to_suffix(transab):
-    return {
-        (True, True): "TT",
-        (True, False): "TN",
-        (False, True): "NT",
-        (False, False): "NN",
-    }[tuple(transab)]
 
 
 def _test_gemm(func, dtype: str, m: int, n: int, k: int, transa=False, transb=False):
@@ -82,47 +73,17 @@ def _test_gemm(func, dtype: str, m: int, n: int, k: int, transa=False, transb=Fa
 
 dtypes = ["float32", "float16"]
 all_transabs = list(product([True, False], repeat=2))
-all_basic_sizes = list(product([1, 3, 4, 16, 127, 128, 129, 133, 1024], repeat=3))
-
-
-def get_bert_sizes(full=True):
-    bert_base_sizes = [
-        # m, n, k
-        (384, 768, 768),
-        (384, 768, 768 * 3),
-        (384, 768, 768 * 4),
-        (384, 768 * 4, 768),
-        (384, 1024, 1024),
-        (384, 1024, 1024 * 3),
-        (384, 1024, 1024 * 4),
-        (384, 1024 * 4, 1024),
-    ]
-
-    # we then multiply m with the batch size
-    if full:
-        batch_sizes = [1, 64]
-    else:
-        batch_sizes = [1]
-    bert_sizes = []
-    for bsz in batch_sizes:
-        bert_sizes.extend([(m * bsz, n, k) for m, n, k in bert_base_sizes])
-    return bert_sizes
 
 
 @pytest.mark.parametrize("dtype", dtypes)
-@pytest.mark.parametrize("size", all_basic_sizes + get_bert_sizes(full=False))
+@pytest.mark.parametrize("size", get_basic_size(full=True) + get_bert_sizes(full=False))
 @pytest.mark.parametrize("transab", all_transabs)
 def test_rocblas_gemm_all_cases(dtype, size, transab):
     _test_gemm(getattr(ke, "RocblasGemm_" + dtype_to_suffix(dtype)), dtype, *size, *transab)
 
 
-# ck has various impls to be tested, use the full basic cases will result too many cases to test.
-# So we use a reduced combination here.
-reduced_basic_sizes = list(product([1, 4, 127, 133], [3, 16, 128], [3, 129, 1024]))
-
-
 @pytest.mark.parametrize("dtype", dtypes)
-@pytest.mark.parametrize("size", reduced_basic_sizes + get_bert_sizes(full=False))
+@pytest.mark.parametrize("size", get_basic_size(full=False) + get_bert_sizes(full=False))
 @pytest.mark.parametrize("transab", all_transabs)
 def test_ck_gemm_bert_cases(dtype, size, transab):
     wrapper_name = "CKGemm_{}_{}".format(dtype_to_suffix(dtype), transab_to_suffix(transab))
@@ -131,7 +92,7 @@ def test_ck_gemm_bert_cases(dtype, size, transab):
 
 # Tunable is basically wrapped around of rocblas and ck gemm, so no need for full tests
 @pytest.mark.parametrize("dtype", dtypes)
-@pytest.mark.parametrize("size", reduced_basic_sizes + get_bert_sizes(full=False))
+@pytest.mark.parametrize("size", get_basic_size(full=False) + get_bert_sizes(full=False))
 @pytest.mark.parametrize("transab", all_transabs)
 def test_gemm_tunable_bert_cases(dtype, size, transab):
     wrapper_name = "GemmTunable_{}_{}".format(dtype_to_suffix(dtype), transab_to_suffix(transab))
